refactor ShardingTransactionEvent
