+++ sharding-transaction/src/main/java/io/shardingsphere/transaction/event/ShardingTransactionEvent.java
@@ -17,28 +17,20 @@
 
 package io.shardingsphere.transaction.event;
 
-import com.google.common.base.Optional;
 import io.shardingsphere.core.constant.TCLType;
-import io.shardingsphere.core.exception.ShardingException;
-import lombok.Getter;
 
 /**
- * XA transaction event.
+ * Sharding transaction event.
  *
  * @author zhaojun
+ * @author zhangliang
  */
-@Getter
-public final class XATransactionEvent extends TransactionEvent {
+public interface ShardingTransactionEvent {
     
-    private final String sql;
-    
-    public XATransactionEvent(final TCLType tclType, final String sql) {
-        super(tclType);
-        this.sql = sql;
-    }
-    
-    @Override
-    public Optional<ShardingException> getException() {
-        return Optional.fromNullable((ShardingException) super.getException().orNull());
-    }
+    /**
+     * Get transaction operation type.
+     * 
+     * @return transaction operation type
+     */
+    TCLType getOperationType();
 }
